fix: improve error handling.
When a node connection error happens, we don't want the user to retry, but
instead fail silently and try again. This also the case when
checkWithdraw and checkMint, so after a Near wallet redirect, the url
params should be cleared carefully so that these checks can try again if
something throws. Failed transactions are clearly marked as FAILED and
all other cases should faild silently to re-try at the next checkStatus
loop.